refactor: directories restructure
