+++ .gitignore
@@ -14,3 +14,5 @@ yarn-error.log*
 # Personal
 TO-DO.md
 /.example
+
+cuconfig.json
